Trading/basket/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mrscore"
version = "0.1.0"
description = "Mean reversion scoring engine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2,<3",
  "pyyaml>=6",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
